ITAMHack/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiogram==3.13.1
sqlalchemy==2.0.23
asyncpg==0.29.0
redis[hiredis]==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0

